renamed root namespace and changed project name
Since this solution no longer uses a tray icon I changed the name to BackgroundApp.
