+++ DAWPresenceBackgroundApp/AppConfiguration.cs
