+++ DAWPresenceBackgroundApp/DAWs/BitwigStudio.cs
