+++ DAWPresenceBackgroundApp/DAWs/StudioOne.cs
